requirements.txt

pandas
numpy
yfinance
pyyaml
pyarrow
